+++ src/translations/en-US.instructions.json
@@ -71,7 +71,6 @@
     "dispute.account": "Label for a balance account",
     "dispute.contactSupportToDefendThisDispute": "Label to inform user that it is not possible to defend the dispute via component and they can contact support.",
     "dispute.continue": "Button label to continue uploading defend documents.",
-    "dispute.defendDispute": "Label of an action button that opens a modal to upload some documents to defend the dispute",
     "dispute.defendDisputeTitle": "Button label to start defend dispute flow.",
     "dispute.defendDocumentAcceptableFormats": "Guidance text to inform user dispute defense documents should have only specified file formats",
     "dispute.defendDocumentMustBeInEnglish": "Guidance text to inform user dispute defense documents should be in English",
@@ -105,6 +104,7 @@
     "disputes.chargebacks": "Label referring to chargeback disputes.",
     "disputes.consumerDispute": "Reason indicating the consumer initiated the dispute.",
     "disputes.currency": "Label indicating the currency used in the disputed transaction.",
+    "disputes.defendChargeback": "Label of an action button that opens a modal to upload some documents to defend the dispute",
     "disputes.defendDispute": "Button label allowing users to defend against a dispute.",
     "disputes.defendedOn": "Label showing the date on which the dispute was defended.",
     "disputes.defenseReason": "Label indicating the reason provided when defending a dispute.",
