+++ src/translations/en-US.json
@@ -117,7 +117,6 @@
     "dispute.account": "Account",
     "dispute.contactSupportToDefendThisDispute": "%{contactSupport} to defend this dispute",
     "dispute.continue": "Continue",
-    "dispute.defendDispute": "Defend dispute",
     "dispute.defendDisputeTitle": "Defend dispute",
     "dispute.defendDocumentAcceptableFormats": "JPEG, TIFF, or PDF formats only",
     "dispute.defendDocumentMustBeInEnglish": "Must be in English",
@@ -140,7 +139,7 @@
     "dispute.submit": "Submit",
     "dispute.uploadDefenseDocumentsInformation": "Upload documents that support your dispute defense. Once submitted, no changes can be made.",
     "disputeDetails": "Dispute management",
-    "disputes.accept": "Accept",
+    "disputes.accept": "Accept Chargeback",
     "disputes.acceptDispute": "Accept dispute",
     "disputes.acceptDisputeDisclaimer": "By accepting, you agree that the disputed amount will not be returned to your account and the dispute will be marked as Lost.",
     "disputes.accepted": "Accepted",
@@ -152,6 +151,7 @@
     "disputes.consumerDispute": "Consumer dispute",
     "disputes.contactSupportToDefendThisDispute": "%{contactSupport} to defend this dispute",
     "disputes.currency": "Currency",
+    "disputes.defendChargeback": "Defend chargeback",
     "disputes.defendDispute": "Defend dispute",
     "disputes.defendedOn": "Defended on",
     "disputes.defenseReason": "Defense Reason",
